requirements.txt
selenium
pytest
pytest-html
pytest-xdist
pyclean
python-dotenv